Added coding for Yipetalichthys_Update Character list 2023 to be compiled.md
Added coding for Yipetalichthys and revised some codings for Petalichthyida, focusing on revisions for features that were coded despite not being preserved. The added feature is P5.  548 The stalk-shaped occipital region is equivalent to an elongated occipital region, so there is no need to add an elongated occipital region as a feature.